CI: Run dmesg test
Run a dmesg test to check for presence of kernel log messages of level
"warn" or higher during boot.

Existing kernel log messages have been added to an ignore list, and the
job is marked as allow_failure. In this way, only new kernel log
warnings and errors will be detected by the test, and cause the pipeline
to be marked as "passed with warnings".