[MERGED] arm64: dts: rockchip: improve PHY handling for ROCK 4D
Add information, that the PHY is RTL8211F based. This helps the
system to probe the right driver, so that the clock is enabled.

Also assign 25 MHz as clock rate for REFCLKO25M_GMAC0_OUT, which
is used by some board variants (early prototype boards have a
dedicated crystal instead, but those may also enable this clock).

While at it switch from non-standard "enable-gpio" to "enable-gpios".

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>