net: phy: realtek: Reset after clock enable
On ROCK 4D we are seeing some issues with PHY detection after
it switched from a dedicated crystal to providing the 25 MHz
clock from the SoC.

Let's try to fix this up by doing a reset after enabling the clock
to avoid the PHY seeing any clock glitches. The datasheet does not
specify any power sequencing requirements for the clock and seems
to expect that a dedicated crystal or always-on clock is being used.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>