arm64: dts: rockchip: use downstream PHY timing setup
With the production Rock 4D boards in our LAVA setup, we are
seeing connection issues. After the recent fixes for the PHY
setup itself the link is established properly, but it seems
no data gets through as DHCP does not work. As things are
working in U-Boot, this is obviously a software issue.

Downstream uses different timing setup; try importing that.

Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>